+++ tests/test_exess_chelpg.py
@@ -4,22 +4,7 @@ from pathlib import Path
 from pprint import pp
 
 from rush import exess
-from rush.client import RunOpts, download_object, save_json, set_opts
-
-
-def test_exess_chelpg():
-    set_opts(workspace_dir=Path.cwd() / "test-runs")
-    data_dir = Path(__file__).parent / "data"
-    res = exess.chelpg(
-        data_dir / "tyk2_ejm_31_t.json",
-        run_opts=RunOpts(
-            name="Rush-Py Test EXESS Energy 03: ChelpG",
-            tags=["rush-py", "test", "tyk2+ejm-31"],
-        ),
-        collect=True,
-    )
-    pp(res, width=130, compact=True, stream=sys.stderr)
-    save_json(res, name="tyk2_ejm_31_chelpg_charges.json")
+from rush.client import RunOpts, download_object, set_opts
 
 
 def test_exess_energy_chelpg():